connector/pyproject.toml
[project]
name = "lobster-town"
version = "0.1.0"
description = "🦞 Connect your OpenClaw to Lobster Town"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "websockets>=13.0",
    "pynacl>=1.5.0",
    "rich>=13.9.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
]

[project.scripts]
lobster-town = "lobster_town.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["lobster_town*"]
